Improve structure and documentation of Loading connector
